# SourceCode/swagger.yml
...
openapi: 3.0.0
info:
  title: "Trail microservice API"
  description: "An API for trails"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    LocationPoint:
      type: "object"
      properties:
        LocationPointId:
          type: "integer"
        Latitude:
          type: "string"
        Longitude:
          type: "string"
...
    TrailLocationPoint:
      type: "object"
      properties:
        LocationPointId:
          type: "integer"
        TrailId:
          type: "integer"
        OrderNum:
          type: "integer"
  parameters:
    LocationPointId:
      name: "LocationPointId"
      description: "Id of a location point"
      in: path
      required: True
      schema:
        type: "integer"
...
paths:
...
  /locationpoints:
    get:
      operationId: "locationpoints.getLocationPoints"
      tags:
        - Location points
      summary: "Get all location points"
      responses:
        "200":
          description: "OK"
        "500":
          description: "Server error"

    put:
      operationId: "locationpoints.editLocationPointById"
      tags:
        - Location points
      summary: "Update location point"
      requestBody:
        description: "Location point to update"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pointJSON"
              $ref: "#/components/schemas/LocationPoint"
      responses:
        "201":
          description: "Successfully updated location point"
        "400":
          description: "Bad request"
        "404":
          description: "Location point not found"
        "500":
          description: "Server error"

    post:
      operationId: "locationpoints.addLocationPoint"
      tags:
        - Location points
      summary: "Add new location point"
      requestBody:
        description: "Location point to add"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "pointJSON"
              $ref: "#/components/schemas/LocationPoint"
      responses:
        "201":
          description: "Successfully added location point"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"
  /locationpoints/{LocationPointId}:
    get:
      operationId: "locationpoint.getLocationPointById"
      tags:
        - Location points
      summary: "Gets location point with given Id"
      parameters:
       - $ref: "#/components/parameters/LocationPointId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Location point not found"
        "500":
          description: "Server error"

    delete:
      operationId: "locationpoints.deleteLocationPointById"
      tags:
        - Location points
      summary: "Deletes location point with given id"
      parameters:
       - $ref: "#/components/parameters/LocationPointId"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "404":
          description: "Location point not found"
        "500":
          description: "Server error"

# Location Points
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# Link entities
  /traillocationlink:
    post:
      operationId: "linkhelper.newTrailLocationPoint"
      tags:
        - Link entity
      summary: "Add new trail-location point link"
      requestBody:
        description: "Location point and trail to link"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailLocationPointJSON"
              $ref: "#/components/schemas/TrailLocationPoint"
      responses:
        "201":
          description: "Successfully added trail location point"
        "400":
          description: "Bad request"
        "500":
          description: "Server error"

    delete:
      operationId: "linkhelper.deleteTrailLocationPoint"
      tags:
        - Link entity
      summary: "Delete trail-location point link"
      requestBody:
        description: "Location point and trail to delete link of"
        required: True
        content:
          application/json:
            schema:
              x-body-name: "TrailLocationPointJSON"
              $ref: "#/components/schemas/TrailLocationPoint"
      responses:
        "201":
          description: "Successfully deleted trail location point"
        "400":
          description: "Bad request"
        "404":
          description: "No TrailLocationPoint found"
        "500":
          description: "Server error"
...